pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-automation"
version = "0.1.0"
description = "CV tooling project"
requires-python = ">=3.12"
dependencies = ["jsonschema", "PyYAML>=6.0,<7", "typer>=0.12,<1"]

[project.scripts]
cvtool = "profilecli.cli:app"

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["src"]
